add(gameplay): giant(character) demo including animations, vfx

+++ Assets/_Project/_Scripts/Feature/Animation/Runtime/PlayerAnimationDriver.cs
@@ -197,7 +197,7 @@ namespace CaseStudy.Feature.Animation.Runtime
 
             if (_hasAbilityIndex && _hasPendingAbilityIndex)
             {
-                _ownerAnimator.SetInteger(_abilityIndexHash, _pendingAbilityIndex);
+                _ownerAnimator.SetFloat(_abilityIndexHash, _pendingAbilityIndex);
                 _hasPendingAbilityIndex = false;
             }
         }
